refactor constraints into models
