+++ frex/models/__init__.py
@@ -1,6 +1,4 @@
 from .domain_object import DomainObject
 from .explanation import Explanation
 from .candidate import Candidate
-from .constraint_solution_section import ConstraintSolutionSection
-from .constraint_solution_section_set import ConstraintSolutionSectionSet
-from .constraint_solution import ConstraintSolution
+from .constraints import *
