+++ frex/utils/__init__.py
@@ -1,4 +1,3 @@
-from .constraints import *
 from .vector_similarity_utils import VectorSimilarityUtils
 from .constraint_solver import ConstraintSolver
 from .common import *
